requirements.txt
pygame-ce>=2.4.0
pydantic>=2.5.0
toml>=0.10.2
colorama>=0.4.6
pytest>=7.4.0
